pyproject.toml
[tool.poetry]
name = "rnafold"
version = "0.1.0"
authors = ["Louis Stefanuto <louis.stefanuto@gmail.com>"]
description = ""
readme = "README.md"

[tool.poetry.scripts]
rnafold = "rnafold.main:app"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
typer = "^0.12.3"
kaggle = "^1.6.17"
pandas = "^2.2.3"
tqdm = "^4.67.1"
biopython = "^1.85"
matplotlib = "^3.10.1"
torch = "^2.6.0"
seaborn = "^0.13.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
nbstripout = "^0.7.1"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
pre-commit = "^3.8.0"
ruff = "^0.5.6"
mypy = "^1.11.1"
pytest-cov = "^5.0.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.31"


[tool.poetry.group.rhofold.dependencies]
ml-collections = "^1.0.0"
einops = "^0.8.1"
openmm = "^8.2.0"
simtk = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-include = ["*.ipynb"]  # lint and format jupyer notebooks

[tool.ruff.lint]
extend-select = ["I"]  # sort imports like isort https://github.com/astral-sh/ruff/issues/8926
